% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variousOtherHelperFunctions.R
\name{assessTemporalIndependence}
\alias{assessTemporalIndependence}
\title{Assess temporal independence and generate a record table starting}
\usage{
assessTemporalIndependence(
  intable,
  deltaTimeComparedTo,
  columnOfInterest,
  cameraCol,
  camerasIndependent,
  stationCol,
  minDeltaTime,
  countsName
)
}
\arguments{
\item{intable}{data.frame. Data frame with detections}

\item{deltaTimeComparedTo}{character. For two records to be considered
independent, the second record must be at least \code{minDeltaTime} minutes
after the last independent record of the same species
(\code{"lastIndependentRecord"}), or \code{minDeltaTime} minutes after the
last record (\code{"lastRecord"}).}

\item{columnOfInterest}{Character. Either the name of the species or individual
column ID. The latter to obtain a recordTableIndividual}

\item{cameraCol}{Character. The name of the camera column (e.g. "Camera")}

\item{camerasIndependent}{logical. If \code{TRUE}, species records are
considered independent between cameras at same station.}

\item{stationCol}{character. Name of the camera trap station column. Assumes
"Station" if undefined.}

\item{minDeltaTime}{integer. The minimum specified time difference (in
minutes) between records of the same species, at the same station, that are
to be considered independent.}

\item{countsName}{character. Vector of the metadata tag name (as per image
management/tagging software) containing count/abundance information, or the
vector of the name given to the count/abundance folder position within
\code{directoryInfoNames} (e.g. "Counts").}
}
\description{
from a dataframe ' ' This function is intended for situations where a detection
table has already been generated, for examples a \code{recordTable} ahs been
generated and has been manipolated based on spcific user's need, and there is
the need to assess the
indipendence of the detections
}
