% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderIndices.R
\name{folderIndices}
\alias{folderIndices}
\title{Folder indices}
\usage{
folderIndices(inDir)
}
\arguments{
\item{inDir}{character. Directory filepath containing 'Station' directories.
It must either contain images in species subdirectories (e.g.
"inDir/StationA/SpeciesA") or images with species metadata tags instead of
'Species' directories (e.g. "inDir/StationA"). Only include path up to
root/parent folder in which the 'Station'/'Site' subfolders are located
(e.g. "C:/Documents/CamProject" ...). 'Station' folders would be next in the
file path but are not included when defining \code{inDir}.}
}
\description{
Return the indices of the folder structure to identify information needed to
populate \code{recordTable}'s or \code{recordTableIndividual}'s arguments
}
\details{
Depending on your set up, you may have multiple information stored in your
folder structure. \code{camtrapRdeluze} can now extract information from
multiple folders (i.e. not only Station and Species), but it needs to know
their position in the directory path when using \code{recordTable} or
\code{recordTableIndividual}. This function aims to facilitate the position
of the folders for the following arguments:  \itemize{ \item
stationIDposition \item speciesPosition \item cameraIDposition }

If there are multiple paths within \code{inDir} - as it is likely - the
function will look for  the (first) longest and use that as an example.
}
