% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneRowOneDetection.R
\name{oneRowOneDetection}
\alias{oneRowOneDetection}
\title{Expand a record table to have one row for each detection}
\usage{
oneRowOneDetection(intable, countsName)
}
\value{
record table (data.table)
}
\description{
If a record table has a count column, this functon expand the record table to
have a row for each detection based on the count column.
}
\details{
\strong{NOTE}: This function is currently possibly slow (a few mins) for
recortable that are more than 1E4, so be patient!
}
\examples{
# create a test recordTable
nc<-12
nr<-1E4
m <- matrix(sample(1:100, replace = T, size = nr*nc), ncol = nc)
recTabTest <- data.table(m)
recTabTest[, Count:=sample(1:3, replace = T, size = nr)]

system.time(
  test<-oneRowOneDetection(intable=recTabTest, countsName="Count")
)
}
