% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2long.R
\name{wide2long}
\alias{wide2long}
\title{Convert a record table from a wide to long formatting}
\usage{
wide2long(intable, pattern, valueNames)
}
\arguments{
\item{pattern}{Character. The pattern to identify the columns names that need
to be collated (e.g. c("Species$", "Age$"))}

\item{valuesNames}{Character. The vector of names to apply to the conmbines
variables}
}
\value{
A record table
}
\description{
If there are multiple columns in a record table that provide the same
information, for example when metatags are used to provide informaiton on
multiple individuals in the same image (e.g. Ind1Species, Ind1Age,
Ind2Species, Ind2Age, etc.), the user may want to collate all these in one
column (e.g. Species and Age). This function fecilitate this process and
removes lines where all the information being removed are missing (e.g. when
there are no additional individuals).
}
\details{
If there is need to use both wide2long and oneRowOneDetection, the latter
should be used first.
}
